# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/areas": {
      "post": {
        "description": "Creates an automation by binding an action to a reaction for the authenticated user.",
        "operationId": "AreasController_createArea",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAreaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Area successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or action/reaction not available",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new area automation",
        "tags": [
          "areas"
        ]
      },
      "get": {
        "operationId": "AreasController_getAreas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Areas belonging to the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreasListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List areas created by the authenticated user",
        "tags": [
          "areas"
        ]
      }
    },
    "/areas/{id}": {
      "put": {
        "operationId": "AreasController_toggleArea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifier of the area to update",
            "schema": {
              "example": 101,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAreaStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated area status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for the provided status",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Area not found for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable or disable an area",
        "tags": [
          "areas"
        ]
      },
      "patch": {
        "operationId": "AreasController_updateArea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifier of the area to update",
            "schema": {
              "example": 101,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAreaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Area updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Provided data violated validation rules",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Area not found for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update area metadata or configuration",
        "tags": [
          "areas"
        ]
      },
      "delete": {
        "operationId": "AreasController_deleteArea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifier of the area to delete",
            "schema": {
              "example": 101,
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Area deleted successfully"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Area not found for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an area",
        "tags": [
          "areas"
        ]
      }
    },
...
  },
  "info": {
    "title": "Area API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "example": [
              "le nom ne doit pas être vide",
              "l’identifiant d’action doit être un entier"
            ],
            "description": "Messages de validation détaillés",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Action invalide",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "UnauthorizedResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Non autorisé",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Identifiants invalides",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "CreateAreaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "M’avertir lorsqu’une nouvelle issue est créée",
            "description": "Nom compréhensible donné à l’automatisation"
          },
          "actionId": {
            "type": "number",
            "example": 15,
            "description": "Identifiant de l’action à associer à cette area"
          },
          "reactionId": {
            "type": "number",
            "example": 22,
            "description": "Identifiant de la réaction à associer à cette area"
          },
          "actionConfig": {
            "type": "object",
            "nullable": true,
            "description": "Paramètres optionnels de l’action. Renseigner les champs obligatoires définis par le schéma de l’action.",
            "example": {
              "owner": "open-source-org",
              "repo": "cool-project"
            }
          },
          "reactionConfig": {
            "type": "object",
            "nullable": true,
            "description": "Paramètres optionnels de la réaction. Renseigner les champs obligatoires définis par le schéma de la réaction.",
            "example": {
              "channel": "#alerts"
            }
          },
          "dedupKeyStrategy": {
            "type": "string",
            "nullable": true,
            "description": "Nom de la stratégie utilisée pour dédupliquer les déclenchements (laisser vide pour désactiver).",
            "example": "payload-hash"
          }
        },
        "required": [
          "name",
          "actionId",
          "reactionId"
        ]
      },
      "AreaServiceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "slug": {
            "type": "string",
            "example": "github"
          },
          "name": {
            "type": "string",
            "example": "GitHub"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
      "AreaActionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 15
          },
          "key": {
            "type": "string",
            "example": "issue_created"
          },
          "description": {
            "type": "object",
            "example": "Déclenchée lorsqu’une nouvelle issue est créée sur le dépôt",
            "nullable": true
          },
          "configSchema": {
            "type": "object",
            "description": "Schéma JSON décrivant la configuration attendue par l’action",
            "nullable": true,
            "example": {
              "type": "object",
              "properties": {
                "repository": {
                  "type": "string"
                }
              }
            }
          },
          "service": {
            "$ref": "#/components/schemas/AreaServiceSummaryDto"
          }
        },
        "required": [
          "id",
          "key",
          "description",
          "configSchema",
          "service"
        ]
      },
      "AreaReactionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 22
          },
          "key": {
            "type": "string",
            "example": "send_notification"
          },
          "description": {
            "type": "object",
            "example": "Envoyer une notification sur le salon configuré",
            "nullable": true
          },
          "configSchema": {
            "type": "object",
            "description": "Schéma JSON décrivant la configuration attendue par la réaction",
            "nullable": true,
            "example": {
              "type": "object",
              "properties": {
                "channel": {
                  "type": "string"
                }
              }
            }
          },
          "service": {
            "$ref": "#/components/schemas/AreaServiceSummaryDto"
          }
        },
        "required": [
          "id",
          "key",
          "description",
          "configSchema",
          "service"
        ]
      },
      "AreaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 101
          },
          "name": {
            "type": "string",
            "example": "Notify me of new GitHub issues"
          },
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "actionConfig": {
            "type": "object",
            "description": "Configuration d’exécution utilisée par l’action",
            "nullable": true,
            "example": {
              "owner": "open-source-org",
              "repo": "area-project"
            }
          },
          "reactionConfig": {
            "type": "object",
            "description": "Configuration d’exécution utilisée par la réaction",
            "nullable": true,
            "example": {
              "channel": "#alerts"
            }
          },
          "dedupKeyStrategy": {
            "type": "object",
            "description": "Stratégie utilisée pour générer les clés de déduplication des déclencheurs",
            "nullable": true,
            "example": "payload-hash"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-01-15T08:30:00.000Z",
            "description": "Horodatage de création de l’area (ISO 8601)"
          },
          "action": {
            "$ref": "#/components/schemas/AreaActionSummaryDto"
          },
          "reaction": {
            "$ref": "#/components/schemas/AreaReactionSummaryDto"
          }
        },
        "required": [
          "id",
          "name",
          "enabled",
          "actionConfig",
          "reactionConfig",
          "dedupKeyStrategy",
          "createdAt",
          "action",
          "reaction"
        ]
      },
      "AreasListResponseDto": {
        "type": "object",
        "properties": {
          "areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreaResponseDto"
            }
          }
        },
        "required": [
          "areas"
        ]
      },
      "UpdateAreaStatusDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true,
            "description": "Whether the area should be enabled (executed) or disabled"
          }
        },
        "required": [
          "enabled"
        ]
      },
      "UpdateAreaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "M’avertir lorsqu’une nouvelle issue GitHub est créée",
            "description": "Nouveau nom attribué à l’area"
          },
          "actionId": {
            "type": "number",
            "example": 18,
            "description": "Nouvel identifiant d’action à associer à l’area"
          },
          "reactionId": {
            "type": "number",
            "example": 24,
            "description": "Nouvel identifiant de réaction à associer à l’area"
          },
          "actionConfig": {
            "type": "object",
            "nullable": true,
            "description": "Nouvelle configuration de l’action",
            "example": {
              "repo": "new-repo"
            }
          },
          "reactionConfig": {
            "type": "object",
            "nullable": true,
            "description": "Nouvelle configuration de la réaction",
            "example": {
              "channel": "#support"
            }
          },
          "dedupKeyStrategy": {
            "type": "object",
            "nullable": true,
            "description": "Stratégie de déduplication mise à jour (mettre à null pour la désactiver)",
            "example": "payload-hash"
          }
        }
      },
...
    }
  }
}